requirements.txt
Django==2.2
django-crontab==0.7.1
django-extensions==3.1.3
django-shell-plus==1.1.5
djangorestframework==3.12.4
djangorestframework-csv==2.1.1
jedi==0.18.0
numpy==1.17.4
pandas==1.3.2